requirements.txt
pytest>=3.0.7
requests>=2.25.1
pyyaml==5.4.1